observability-host/analyzer/pyproject.toml
[project]
name = "analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-generativeai>=0.8.4",
    "gunicorn>=23.0.0",
    "requests>=2.32.3",
]

[tool.uv.workspace]
members = ["demo-azure"]
